+++ providers/uninstall.rb
@@ -18,16 +18,15 @@
 #
 
 action :run do
-    #remove redis binaries
-    execute "rm -rf /usr/local/bin/redis*" if ::File.exists?("/usr/local/bin/redis-server")
+  #remove redis binaries
+  execute "rm -rf /usr/local/bin/redis*" if ::File.exists?("/usr/local/bin/redis-server")
 
-    #remove configuration file and init script for servers provided
-    unless new_resource.servers.nil?
-      new_resource.servers.each do |server|
-        server_name = server['name'] || server['port'] 
-        execute "rm -rf /etc/redis/#{server_name}.conf" if ::File.exists?("/etc/redis/#{server_name}.conf")
-        execute "rm -rf /etc/init.d/redis#{server_name}" if ::File.exists?("/etc/init.d/redis#{server_name}")
-        execute "rm -rf /etc/init/redis#{server_name}.conf" if ::File.exists?("/etc/init/redis#{server_name}.conf")
-      end
-    end
+  #remove configuration file and init script for servers provided
+  RedisioHelper.each_server(new_resource.servers) do |server|
+    server_name = RedisioHelper.server_name(server)
+    service_name = RedisioHelper.service_name(server)
+    execute "rm -rf /etc/redis/#{server_name}.conf" if ::File.exists?("/etc/redis/#{server_name}.conf")
+    execute "rm -rf /etc/init.d/#{service_name}" if ::File.exists?("/etc/init.d/#{service_name}")
+    execute "rm -rf /etc/init/#{service_name}.conf" if ::File.exists?("/etc/init/#{service_name}.conf")
+  end
 end